+++ src/Snap.Hutao.Server/Snap.Hutao.Server/Model/Entity/EntityFloor.cs
@@ -18,12 +18,12 @@ public class EntityFloor
     /// </summary>
     [Key]
     [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-    public int PrimaryId { get; set; }
+    public long PrimaryId { get; set; }
 
     /// <summary>
     /// 外键
     /// </summary>
-    public int SpiralAbyssId { get; set; }
+    public long SpiralAbyssId { get; set; }
 
     /// <summary>
     /// 引用深渊信息
